scripts/requirements.txt
yt-dlp>=2023.12.30
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
